move applyDarkStyles to utils
